# swagger.yaml
openapi: "3.0.0"

info:
  title: Learning Express and Swagger - Manthan Gupta
  description: A Segment aboout learning and writing docs
  version: 1.1.0
  contact:
    email: guptamanthan280@gmail.com

# servers:
#   - url : "https://localhost:4000/api/v1"
#     description: for localhost server - secure
#   - url : "http://localhost:4000/api/v1"
#     description: for localhost server - regular  

servers:
  - url : "{protocols}://localhost:4000/api/{version}"
    description: for local host
    variables:
      version:
        enum:
        - v1
        - v2
        default: v1
      protocols:
        enum:
        - https
        - http
        default: http

components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: token
    BearerAuth:
      type: http
      scheme: bearer


paths:
  /lco:
    get:
      tags:
      - Strings
      summary : Returns a greet message from LCO
      responses:
        200:
          description: Success
          content:
            application/json:
              schema:
                type: string
                example: "my string"
        400:
          description: Bad Request
        500:
          description: Internal Error         

  /lcoobject:
    get:
      tags:
      - Object
      summary: Returns a unique course
      responses:
        200:
          description: success
          content:
            application/json:
              schema:
                type: object
                properties:
                  id : 
                    type: string
                  name:
                    type: string
                  price:
                   type: integer
        400:
          description: bad request
        500:
          description: Internal server error
 
  /courses:
    get:
      tags:
      - Array
      summary: Returns all courses
      responses:
        200:
          description: success
          content:
            application/json:
              schema:
                type: array
                items: 
                  type: object
                  properties:
                    id : 
                      type: string
                    name:
                      type: string
                    price:
                      type: integer
        400:
          description: bad request
        500:
          description: Internal server error

  /mycourse/{courseId}:
    get:
      tags:
        - Strings
      summary: Returns course that is passed on the parameter
      parameters:
        - name : courseId
          in: path
          required: true
          default: 22
          schema: 
            type: integer
      responses:
        200:
          description: success
          content:
            application/json:
              schema:
                  type: object
                  properties:
                    id : 
                      type: string
                    name:
                      type: string
                    price:
                      type: integer
        400:
          description: bad request
        500:
          description: Internal server error

  /addCourse:
    post:
      tags:
        - Strings
      summary: Adds a new course to existing courses
      consumes:
        - application/json
      produces:
       - applications/json
      requestBody:
        required: true
        content:
          application/json :
            schema:
              type: object
              properties: 
                id:
                  type: string
                name:
                  type: string
                price:
                  type: integer
      responses:
        200:
          description: success
          content:
            application/json:
              schema:
                  type: boolean
        400:
          description: bad request
        500:
          description: Internal server error

  /coursequery:
    get:
      tags:
        - Strings
      summary: learning about query
      parameters:
      - name : location
        in: query
        required: true
        schema:
          type: string
          enum: [delhi, gurgaon , kanpur]
      - name : device
        in: query
        required: true
        schema:
          type: string
          enum: [web, mobile]
      responses:
        200:
          description: success
          content:
            application/json:
              schema:
                  type: object
                  properties:
                    location:
                      type: string
                    device:
                      type: string
        400:
          description: bad request
        500:
          description: Internal server error

        
  /courseupload:
    post:
      tags:
        - Strings
      summary: uploading course image
      parameters:
      - in : headers
        name: auth
      requestBody:
        content:
          multipart/form-data :
            schema:
              type: object
              properties: 
                file:
                  type: string
                  format: binary
      responses:
        200:
          description: success
        400:
          description: bad request
        500:
          description: Internal server error